requirements.txt
psycopg2-binary==2.9.9
Flask==3.0.3
python-dotenv==1.0.0
gunicorn==22.0.0